refactor: Refactor folder path
