+++ app/business/service/invoice/revenue.query.ts
@@ -1,7 +1,7 @@
 // RSC를 위한 data fetch 함수 작성
 // 이 sample 코드는 invoice 도메인에 revenue 도메인 객체(애그리게이트) 쉽게 말하면 걍 데이터를 fetch하는 함수들을 작성
 
-import { API_PATH } from '../api-path';
+import { API_PATH } from '../../api-path';
 
 export type Revenue = {
   month: string;
