+++ app/business/service/result/result.type.ts
