+++ app/business/service/user/user.command.ts
@@ -1,7 +1,7 @@
 'use server';
 
 import { FormState } from '@/app/ui/view/molecule/form/form-root';
-import { API_PATH } from '../api-path';
+import { API_PATH } from '../../api-path';
 import { SignUpRequestBody, SignInRequestBody, ValidateTokenResponse, UserDeleteRequestBody } from './user.type';
 import { httpErrorHandler } from '@/app/utils/http/http-error-handler';
 import { BadRequestError } from '@/app/utils/http/http-error';
