+++ app/business/service/user/user.query.ts
@@ -1,4 +1,4 @@
-import { API_PATH } from '../api-path';
+import { API_PATH } from '../../api-path';
 import { InitUserInfoResponse, UserInfoResponse } from './user.type';
 import axios from 'axios';
 import { getToken } from '../auth';
