+++ app/ui/result/result-category-detail/result-category-detail.tsx
@@ -1,11 +1,11 @@
 'use client';
 import ResultCategoryDetailContent from '@/app/ui/result/result-category-detail-content/result-category-detail-content';
-import { fetchResultCategoryDetailInfo } from '@/app/business/result/result.query';
+import { fetchResultCategoryDetailInfo } from '@/app/business/service/result/result.query';
 import ResultCategoryDetailContentSkeleton from '@/app/ui/result/result-category-detail-content/result-category-detail-content.skeleton';
 import ResultCategoryDetailDialog from './result-category-detail-dialog';
 import { Suspense } from 'react';
 import { QUERY_KEY } from '@/app/utils/query/react-query-key';
-import { ResultCategoryDetailResponse } from '@/app/business/result/result.type';
+import { ResultCategoryDetailResponse } from '@/app/business/service/result/result.type';
 import { useSuspenseQuery } from '@tanstack/react-query';
 
 export default function ResultCategoryDetail({ category }: { category: string }) {
