+++ app/ui/user/user-info-navigator/sign-button-group.tsx
@@ -1,4 +1,4 @@
-import { auth } from '@/app/business/auth';
+import { auth } from '@/app/business/service/auth';
 import SignOutButton from './sign-out-button';
 import UserDeleteButton from './user-delete-button';
 import SignInLinkButton from './sign-in-link-button';
